refactor: move snowflake.ts to utils
